[#42131657] Update lock path for Puppet 3.x
Puppet 3.x now uses separate lock files for "running" and "disabled":
http://docs.puppetlabs.com/puppet/3/reference/whats_new.html#changes-to-agent-lockfile-behavior

I did test using `puppet agent --configprint` to fetch the value. However it
doubles the runtime for this simple operation. Similarly, I tried
maintaining compat with Puppet 2.7, but it was slow and a terrible mess of
conditional Bash.

This will break compat with Puppet 2.7 -- which should only be in
staging/prod for the next couple of days. Nagios will however continue to
report freshness checks.