It was taking too long to turn it off & on again
The graceful stop waits for running processes to finish but when we have an
alert where we have too many connections this is not ideal - we just want to
kill it quickly.

This commit adds a task to kill it and makes that what is called in the hello_it
(have you tried turning it off and on again?) task.