[#43121095] Add puppet.loadhosts task
Loads hosts that use given Puppet class(es). This differs from the `class` task
which only returns nodes of a given "role". It uses the functionality added
to `govuk_node_list` in:

https://github.gds/gds/puppet/commit/d986313df235af0ac225f8e7f3e5838574a3dec0

For example, to run a command on machines that include the `redis` and
`govuk::apps::govuk_crawler_worker` hosts. Note there is `logs-redis-N` in
addition to `redis-N`:

    (fabric)➜  fabric-scripts git:(puppet_loadhosts_by_classes) fab --hide=running,status preview puppet.loadhosts:redis,govuk::apps::govuk_crawler_worker hosts
    logs-redis-1.management.production
    logs-redis-2.management.production
    mirrorer-1.management.production
    redis-1.backend.production
    redis-2.backend.production