Make safe reboots task actually work
You can't just extend env.hosts, because fab tasks are run once for each
host in the host list, and you're trying to extend a list while you're
in the middle of iterating over it.

There are a number of possible solutions.  Here I'm using `execute` to
dynamically change the host list [1].

Another option would be to separate selecting hosts from running tasks
on them, so you'd end up with an interface like this:

    fab staging badger.safe_boxes:2 vm.reboot

which perhaps has an advantage that it allows you to perform other tasks
on this group of machines?  In this case I'm not sure we want that.

[1]: http://docs.fabfile.org/en/1.7/usage/execution.html#dynamic-hosts