Fix rebooting issue
There was an issue where if any one host had packages requiring a
reboot, it would cause fabric to run force_reboot on every host in
env.hosts -- ie if any host needed rebooting, they would all get
rebooted.

This fixes the issue by ensuring that when we execute force_reboot, we
ensure we're only doing so for the current host.