Parallelise kernel maintenance jobs
This job takes quite a while, and will get longer as the number of machines
we use grows in time. To try and speed it up, but also retain some sanity,
add some paralellisation.

As there are a large number of hosts, let's not overwhelm local machine by
running too many concurrent Fabric processes in separate threads. Do this by
using a moving bubble as a kwarg to limit Fabric to a set number of
concurrently active processes.

We could have specified this globally, but doing it this way requires no
prior knowledge of this caveat.