Icinga downtime for services as well as hosts
Make sure that all services associated with a host are also scheduled for
downtime. Otherwise Blinken gets fill with alerts for the services when you
reboot a machine. There's not an easy way to propagate the comment from the
host, but I think it should be clear enough as it is.