Fix Elasticsearch settings on newer versions
Newer versions of Elasticsearch do not return a key of "ok" in the
response body when changing a setting, which means this task
fails on our newest machines.

Amend the conditional so that we only check for "acknowledged",
unless we're on an old version of Elasticsearch in which case
also check for "ok".

We can remove the 2nd conditional once all of the old Elasticsearch
is gone.