Adds recover_origin task to Fabric
Whilst we can fail the site to mirrors easily using Fabric, we can't restore
as easily - at least, it's not a super-quick Fabric script.

This makes it a nice quick fix. The logic is as follows:

  - Re-enable Puppet on the affected class
  - Run Puppet, with some degree of verbosity for confidence
   - This should restore the www.gov.uk Nginx vhost
     - This should restart Nginx because the configuration has changed
  - Prints a success message

Changes the fail_to_mirror success message to point to using this task in
future to recover back to origin from the mirrors.