Delete default vhost as well as www.gov.uk
Delete the default Nginx virtual host as well as the `www.gov.uk`
virtual host.

The default virtual host references the `varnish` upstream which is
defined in the `www.gov.uk` virtual host, meaning that nginx fails to
start if the `www.gov.uk` is not present because it can't find the
`varnish` backend.

This means that the Fabric task `incident.fail_to_mirror` only stops
Nginx on the first cache node; Fabric will not continue to the other
cache nodes because `sudo service nginx start` returns a non-zero exit
code.

It looks like this has been broken since
https://github.gds/gds/puppet/commit/baef41223501337005f1efa4342e01fffaa0139c.