Stop Nginx to fail to mirror
...rather than deleting a virtual host and restarting nginx.

We made this task more selective in 338b449 to prevent the Assets
virtual host from being disabled [1]. Since then, we now support
failover for `assets.digital.cabinet-office.gov.uk` [2], so this is no
longer an issue.

Prefer stopping nginx entirely as it's the simplest thing to do and
means we're not coupling this task to our Nginx virtual host
configuration.

[1]: https://github.com/alphagov/fabric-scripts/pull/21#issuecomment-31009889
[2]: https://github.gds/gds/cdn-configs/commit/8ca1877f55806f6d2f5ce3fb0d95b0142b7fa503#diff-e2f1431e532ca56328341b841c0cdaaf