Add a postgres sync command
This relies on agent forwarding to save copying a temporary file down to
where the script is running and back up again. This is temporarily
enabled with a fabric command switch.

This runs pg_dump to dump the database on the source machine and pipes
it to postgres running on the destination machine - as the two sides of
the pipeline run as different users fabrics sudo decorator cannot be
used. pg_restore is used to restore the dump file as this is more
robust when syncing between different postgres versions