Make wait_for_status more robust and quiet
Under some conditions (eg. shortly after a reboot) fabric will prompt
for a password when trying to ssh onto a machine. The abort_on_prompts
setting causes this to raise an exception rather than just hang. This
means that the tasks can be called sooner and slow starts do not cause a
safe reboot to hang. It's a bit nasty because abort_on_prompts raises a
SystemError which needs to be caught explicitly due to not inheriting
from Exception. This is likely to change in Fabric 2.

This change also hides a lot of wait_for_status output. Given that it
tends to run for quite a few iterations having lots of output can make
it hard to spot the important details.