Fix nagios_hostname validity checks for new naming
Since renaming all the things for the Carrenza
migration, we need to adjust these validity checks
against the `host`. Staging/Preview environments
when split on ‘.’ have 7 elements as they contain
the environment in the FQDN, whereas production
does not contain the environment in the FQDN and
hence has only 6 elements.