Automate emergency publishing to frontend + static
To publish emergency content we need frontend + static
applications checked out locally, make changes, commit,
wait for CI to build, merge, deploy etc.

In the case of an emergency this is going to take too
long.

This task builds upon prior art to automate all of the
steps in the opsmanual (soon to be updated).

See opsmanual for more info on the available tasks.