Allow specifying multiple classes at the same time
This was previously possible by doing:

```
fab production class:cache class:bouncer do:uptime
```

But that would run the `class` task for bouncer on each of the cache
machines (which is not a problem, just noisy).

This commit changes the parameter to the `class` task so that
multiple arguments can be passed in.