Improve `nagios.loadhosts`
My typical use case is searching for unhandled problems in Icinga
and running a task on all machines which match.

Instead of having to find an Icinga URL that outputs JSON it would
be more useful for me to only provide a search string.

This commit changes from using the `requests` library locally to running
curl on the remote machine (because that's the only thing that's installed
there).

We provide the `--insecure` flag to curl because the SSL certificate
for the monitoring machines doesn't have `alert.cluster` as a valid
hostname.

I've also added the `@hosts` decorator to this task so that a machine
to run it on doesn't have to be provided.