Fix loadhosts task in preview
Preview monitoring is behind basic auth, so in order to load hosts in
that environment we have to be able to pass in a username and password.

This checks the status code of the response and prompts for a password
if it's a 401.

It does mean that we're running the same HTTP request twice which isn't
ideal but also shouldn't be too much of a problem.

This uses some super funky double curly braces for `{{http_code}}`
to stop Python from interpreting the single curly brace as a variable
interpolation in a string which has `format()` called on it.