Add a task for purging everything from our caches
This is needed so that emergency publishing notices (which should appear
on every page) will quickly and consistently appear across the site at
the same time.

We use bans as the method to invalidate Varnish objects as it allows us
to invalidate all objects matching a regular expression, which in our
case matches all current objects.