Clear static generated templates for emergency publishing
Our various frontend applications use the wrapper.html.erb and
header_footer_only.html.erb templates. They get these templates using
the Slimmer gem. And this gem fetches these templates from the static
application, located using the ASSET_ROOT.

When static is deployed there are no generated wrapper.html.erb or
header_footer_only.html.erb templates. At the first request of either of
these, the application will generate the template. The template will be
placed in the public/template directory. From that point on, the
templates are served by nginx.

This function clears the public/template directory to force it to be
regenerated to include the emergency campaign.

It's been created on the back of a fake incident we ran on 2nd line to
run emergency publishing.