Also check 302s when replaying mapit traffic
Mapit will take a url like `/area/<gss-or-ons-code>` and redirect to `/area/<internal-id>` so we want to make sure we still get a 302 for these.  We also leave an explanation for any 404s you may get when replaying the 200s because a new database may change internal ids meaning the `/area/<gss-or-ons-code` url will still 302, but the old `/area/<internal-id>` that would have resulted is now a 404.

Replaying the traffic isn't a perfect test as we're not catering for the above, nor do we check that the 200s have the same (or same enough for the parts of the response that we care about) response body.