Provide new version of update database script
On new mapit servers (mapit-*.api) the responsibility for populating the db from a new database dump has moved from puppet to a script deployed with the application.  The new update_database_via_app task handles invoking this script and orchestrating stopping / restarting services that run the app (nginx, memcached, etc...).

Note that the application script will run commands that `sudo -u postgres` because it has to .  Normally we'd run application scripts as deploy, but deploy can't passwordless sudo, so we just run the command as the current user who should be able to do this.

We retain the previous version of the task as update_database_via_puppet so we can still update the old servers (mapit-server-*.backend) on production.