Disable puppet around the work we do
There's a chance that a scheduled puppet run could kick-off while our tasks are executing.  This would mean that services we expect to be down for the duration could be brought back up before we are ready and the app would start receiving traffic while the db is empty, or being updated.  This could have problematic results.