Don't use puppet to bring services back up
In the task that uses the app script to update the database we don't want to use
puppet to bring back up the services we stop.  We need the services to come back
up in a specific order (the reverse of the order we stopped them in).  This
would require specifying ordering in puppet for the mapit app which would make it less like a normal govuk app so we decided to keep the explicit ordering local to the script instead.