Clear all static layouts when flushing cache
As part of emergency publishing we update static layout templates,
which are page cached (to  public, from where they're then served
by nginx, not rails).

We manually clear two of the layout files, but not `core_layout`
which was more recently introduced.

This is a pretty short term fix, and will break again if we add or
rename a layout.

A medium term solution is moving the uncaching logic to a static
rake task, so that lives in the same place as the templates, and
is less likely to fall out of sync.

A longer term solution is using ESI, and side-stepping cachign
entirely.

As we've not gone for a medium or long term solution yet, this
fixes the issue causing apps to not update after a emergency
publishing/unpublishing