Reload app before clearing static files
In the case where we reload the app after clearing the files, there is
a race condition where the old workers could regenerate the files
after the files have been removed.

When we reload the app first, there is still a race condition where the
unicorn workers which are waiting to be killed by unicornherder could
regenerate the files. But it's probably a better race condition.