Restart frontend apps rather than reloading
Reloading applications uses unicornherder to start new processes
and give them time to warm up before killing the old ones.

However it introduces a race condition in the generation of static's
file cache where we have both old and new workers running.

When we're publishing an emergency notification on GOV.UK it's simpler
to just do a hard restart of the application. This might cause requests
to be dropped for a very short amount of time (but no more than a normal
application deploy would).