Task to respawn large unicorn workers
In lieu of an automated Icinga event handler to solve issues with high
memory for certain apps, this script will gracefully kill the largest
unicorn worker for a specified app.

It issues a SIGQUIT to the worker which will tell it to finish serving
the current request then exit. The master will notice this and fork a
new worker with less bloat.