Remove unsupported/unused commands for apt
We don't use these two apt commands and we shouldn't have to run them
manually, so remove them.

`upgrade()` is potentially bad because it can create disparity between
machines or environments. We automatically upgrade Ubuntu packages with
security updates using `unattended-upgrades` and we should use Puppet to
upgrade the packages we manage by setting `ensure` to `latest` where
appropriate.

`dist_upgrade()` is potentially bad for the same reasons, but also
because it may remove packages. There's no need for us to run it
manually.