Make test command usable on dev environment
When running `curl -sI http://localhost:3108/...` on dev, the HTTP header in
the response is HTTP/1.0.  Although this is script can't be directly in the
dev VM, it's a good resource of how to test MapIt locally prior to merging
changes or uploading a new database to other environments.

This change makes the script slightly more ambivalent about the HTTP version.