Test MapIt with unique urls
When we deploy a new version of MapIt or a new database, we replay the previous
day's traffic against it to ensure that we can serve all the requests as least
as well as before.

It seems to add little value to request the same thing multiple times, and just
takes longer.  This will get the unique URLs instead of all of them.

To give some context about the scale of the problem, the last set of logs we
used for this process had 35,000 HTTP 200 responses in the day.  This would
have reduced to 18,000 unique requests, which still takes a little while to
replay, but half as log as before.