Add script to output puppet status in environment
By running:

```
GITHUB_ACCESS_TOKEN=<TOKEN HERE> ./bin/puppet_versions.sh
```

We get an overview of what puppet version is running where. The script requests
for an environment, and then runs a fab script in all hosts in order to
determine which version of puppet is on each server and the last time puppet
ran.

NOTE: teh fab script runs in parallel using 10 processes. This means it's fast,
but it also means sometimes not all output is received in a way the script
expects. This can be solved by re-running it. This isn't ideal, but most of the
times the script runs well.

The output is something like this:

```
Enter the environment: integration

==> The following releases are live in integration

release_18123
release_18122

==> Puppet details per server

api-elasticsearch-2.api.integration
    Current release: release_18123
    Last puppet run: Tue Dec  6 15:01:18 2016
draft-content-store-2.api.integration
    Current release: release_18123
    Last puppet run: Tue Dec  6 14:42:06 2016
logs-redis-1.management.integration
    Current release: release_18123
    Last puppet run: Tue Dec  6 15:08:55 2016
monitoring-1.management.integration
    Current release: release_18123
    Last puppet run: Tue Dec  6 14:49:35 2016
...
```