Add fabric task to clear cached templates
The templates in static and frontend are cached in the application.

When the emergency banner is deployed (and removed) we need to purge
these caches so that static will regenerate the templates.

There is already a fabric function that clears down the wrapper
templates that display the banner on most pages. We need to add the
homepage template to this list.

We also need to purge tmp/cache in the frontend app.

Set the `warn_only` flag to True so that fabric won't throw an error
and quit if the template it's trying to purge has already been removed.

This fabric script is a temporary work around until centralised
caching is implemented.