Add task to blank bundler config
If the gemstash server fails then bundler won't automatically failover
(due to the age of some versions of bundler). In order to continue
deploiying if the gemstash server is unavailable (and can't be brought
back up) we can remove the configuration that says to use the server and
return to using the default values (i.e. rubygems).

This adds a task that will lock puppet on the targeted machine then
empty the bundler config file.