Add fab script to purge all attachments for a given document
This is a short term 'fix' for an ongoing problem. There is an
intermittent issue where content-type for an attachment is being dropped
somewhere in our stack. While this is being traced this script should
allow the purging of all attachments on a given document that has been
reported as having issues.

See these Zendesk tickets for further context:

https://govuk.zendesk.com/agent/tickets/2498475
https://govuk.zendesk.com/agent/tickets/2530401