Add ability to set and remove app environment varibles
This is useful when testing new environments variables without
having to deploy a run puppet. We're also going to be using these
commands while migrating the Email Alert API from GovDelivery to
GOV.UK Notify.

The command won't restart the app as we don't know all the
possible services that might need to be restarted, so that's
down to the user of the command.