Add task to truncate tables in Email Alert Api
As part of the switch over to using Notify, we want to truncate the new
tables in Email Alert Api that currently contain only test data. There
is a rake task to do this in Email Alert Api, but to ease the deploy, we
want a fab task to call the rake task so we don't have to ssh or rely on
Jenkins.