Speedup fabfile by lazily fetching hosts/classes
The previous fabfile was particularly slow for AWS environments, where
it would sequentially fetch the hosts for each class. This has been
replaced with explicit calls to fetch a group of hosts as they are
needed. The 'numbered' task is not AWS-compatible and was removed.

Because the classes are no longer pre-fetched, it's now necessary to
specify 'localhost' in order to fetch them - they can only be fetched
from within the infrastructure. As the 'hosts' decorator conflicts with
the 'hosts' task, the 'classes' task needed moving up to make it work.