Support staging_aws and production_aws
I seem to always forget where the `aws` goes (and always gets it wrong
the first time 🙄). I think this is because `govukcli set-context` has
the `aws` at the end which is opposite to the ordering of the current
Fabric tasks.

This PR allows you to type in both ways.