Check nginx config before gracefulrestart
The name `nginx.gracefulrestart` implies two things: nginx will be
restarted, and it will be graceful. However, if the nginx config file is
invalid, this task will still stop nginx, nicely draining all the
connections, but then won't be able to start it back up again.

`configtest` simply runs `service nginx configtest` to test the
configuration. This is now a step to run in the `gracefulrestart` task
- hopefully preventing nginx from being shut down if it won't be able to
start back up!