Remove the statsd.create_counter task
I don't think this has been used in a while, and I also think there
are better approaches now of addressing the problem it's trying to
solve.

I think this was about getting rid of UNKNOWN Icinga checks, but it's
better just to acknowledge the failing check, or change the Graphite
query if the data just hasn't arrived yet.