Change Content Publisher Maintenance Mode
The Developers have said it would be better to put Content Publisher in
maintenance mode during the migration rather than have it in a GET only mode.
It is different from Whitehall in the sense that other APIs do not access the
content/assets created by it. Content Publisher is instead only accessed
through it's user interface.